Restore marking files to default
This commit configures the marking files for the first of two builds
required to bootstrap the model compiler to migrate from the existing
persistence mechanism, which supports only formalized associations to
the new one provided by #12020, which supports unformalized and
formalized associations using a new LINK SQL statement.
It also removes a .classpath file that was mistakenly committed.

To build a version of Ciera supporting persistence of unformalized
associations using a version that does not contain this support, follow
these steps:
- Merge 12020 into the branch you will build.
- Build Ciera version X using the existing markings and the existing
version of Ciera.
- Change ciera/src/tool-core/gen/application.mark:
-- Remove: *,NonPersistentInstanceIds,*,true
-- Add: io.ciera.tool::core,NoInstanceIdsForLoad,Component
- Change ciera/src/tool-sql/application.mark
-- Remove: *,NonPersistentInstanceIds,*,true
- Change ciera/src/tool-template/application.mark
-- Remove: *,NonPersistentInstanceIds,*,true
- Build Ciera version Y using the updated markings and version X of
Ciera.  Note that version Y should be built such that tool-core,
tool-sql, and tool-template are dependent on version Y of the runtime
since version X of the compiler is an intermediate version of no use for
building “normal” applications.  Therefore version X of the runtime will
not be published and available for applications being built with version
Y, so we must avoid having tool-core, tool-sql, and tool-template
dependent on version X of the runtime.

Version Y may now be used to build any “normal” application without
employing either of the special marks, NonPersistentInstanceIDs and
NoInstanceIdsForLoad.

Version Y may also be used to build Ciera using the 2.4.0 markings,
which employ the NonPersistentInstanceIDs mark.  This will work fine so
long as all the models used to build Ciera use formalized associations.

Version Y may also be used to build Ciera using updated markings that
eliminate the use of NonPersistentInstanceIDs while employing
NoInstanceIdsForLoad in ciera/src/tool-core/gen/application.mark.  This
approach enables the removal of formalization from the models used to
build Ciera.