Added ClassInComponent
Parse-all identifies 21 errors, all of which should be rectified by
adjusting relevant actions that establish, traverse, or remove links on
R408, which was a simple 1:M and is now a M:M:1 associative association.
When establishing a link, an instance of ClassInComponent must first be
created and then employed with the 'using' clause of the 'relate'
statement.
Removing a link also requires the 'using' clause in the 'unrelate'
statement followed by the deletion of the ClassInComponent instance.
Each traversal of R408 requires analysis to determine what sort of
adjustment is necessary.