Interim save
Templates and render operations still need work to generate a
component-independent interface for a modeled class only when it is
defined outside a component.  When a modeled class is defined within a
component, the generated code for it should be identical to that
generate by the existing version of the compiler.