Commit after unplanned shutdown
I was not yet ready to commit these changes, but my laptop shut itself
down while I was not looking, and that rarely goes well with Linux
running under VMWare.
So, the xtUML editor was having trouble displaying some of the functions
I was working on.  May need to roll back to the previous commit.