Fixed relate sequence issue
The addition of ComponentSpecificType requires some care in the
sequencing of relating instances in the hierarchy.
A newly created instance of a subtype of Type cannot be linked to a
newly created instance of Type until after returning from the invocation
of resolve().  And, because the associations are formalized, the relate
statements must execute after the identifying attributes for Type have
been set, and they must be executed in descending order down the
super/subtype hierarchy.