Revert "move quoting of literal strings to translation phase"
This reverts commit 5c5f1019c1ceb5cbe2863c743547569d98e7a381.

The commit was originally made to address the `endl` literal in MASL. This
should probably be translated as a named reference and not an actual string
literal.  I think now it is wrong to put the quotes in the translation phase
since the target should determine how the value is quoted. I may change my mind
again. Either way, the quotes are not a valuable addition to the data so it
makes sense to only deal with them in one place (render).

This change did cause a problem which precipitated the reversion. The template
tool generator began to fail in its dealings with string literals.