Fix issue with relate/unrelate using
The link instance expression is evaluated twice. This is actually not an
issue at all with OAL since the link instance expression must be a
simple variable reference, however MASL allows the link instnace
expression to be any expression of the correct instance type. Because of
this, an expression which has side effects can be used (e.g. create
instance expression or service invocation). Obviously such an expression
cannot be allowed to be evaluated twice, so this change makes it so that
it is evaluated once and stored in a local variable which is passed to
the relate and unrelate calls.